refactor(api): move TargetProfileForAdmin into prescription
