+++ api/src/prescription/target-profile/domain/models/TargetProfileForAdmin.js
@@ -1,6 +1,6 @@
-import { DomainError } from '../errors.js';
-import { AreaForAdmin } from './index.js';
-import { categories } from './TargetProfile.js';
+import { DomainError } from '../../../../shared/domain/errors.js';
+import { AreaForAdmin } from '../../../../shared/domain/models/index.js';
+import { categories } from '../../../../shared/domain/models/TargetProfile.js';
 
 class TargetProfileForAdmin {
   constructor({
